refactor(app): store folder restructure + behaviorist sweep + page architecture
Apply audit follow-ups across stores, components, pages:

Store folder restructure
- Split src/store/ into sidebarStore/ and toastStore/ subfolders
- Make slices private to their folder; expose only public types via barrel
  (sidebarView.ts, newPipelineForm.ts for shared schema/enum)
- Remove src/store/index.ts root barrel; consumers import from
  @/store/sidebarStore or @/store/toastStore

Behaviorist naming
- sidebarSlice: delete unused setNewPipelineOpen; rename setView/setSearchOpen
  and handle{Show,Open}* to UI-event names
  (handleSidebarLocationChange, handleSearchDialogOpenChange,
   handleSidebarMainViewButtonClick, etc.)
- uiSlice: setNodeLlmContent → applyNodeLlmContent (data sink, not UI handler)

Component splits (one-component-per-file)
- InputSnapshotPanel: 5 components → 5 files + helpers
- AgentRunsPanel: 8 components → 8 files + helpers

Page architecture (Wrapper/Content split)
- DistillationsPage + DistillationDetailPage move all UI/data fetching into
  *Content/ subfolders; Wrapper becomes thin

Zod-infer-type
- Delete pages/CanvasPage/RunConsole/types.ts; reroute Job/JobStatus imports
  to @repo/schemas

UI components
- DistillationActionBar replaces native <select> with @repo/ui Select;
  matching slice handler signature for onValueChange (string | null)

Barrel exports
- 3 page-level index.ts: export { Name } → export *
- NodeCard: stop re-exporting NodeTheme from non-index file;
  expose via NodeCard/index.ts instead

tsc, oxlint, vitest (319/321) all green.
