+++ apps/app/src/pages/CanvasPage/NodeCard/index.ts
@@ -1,2 +1,3 @@
 export * from "./NodeCard";
+export * from "./nodeCardTheme";
 export * from "./nodePorts";
