+++ apps/app/src/pages/DistillationDetailPage/DistillationDetailPageContent/index.ts
@@ -1,0 +1,1 @@
+export * from "./DistillationDetailPageContent";
