+++ apps/app/src/pages/DistillationsPage/DistillationsPageContent/index.ts
@@ -1,0 +1,1 @@
+export * from "./DistillationsPageContent";
