+++ apps/app/src/pages/OperationsPage/OperationCard/index.ts
@@ -1,1 +1,1 @@
-export { OperationCard } from "./OperationCard";
+export * from "./OperationCard";
