+++ apps/app/src/pages/OperationsPage/OperationListRow/index.ts
@@ -1,1 +1,1 @@
-export { OperationListRow } from "./OperationListRow";
+export * from "./OperationListRow";
