+++ apps/app/src/store/sidebarStore/newPipelineForm.ts
@@ -1,0 +1,8 @@
+import { z } from "zod/v4";
+
+export const newPipelineFormSchema = z.object({
+  name: z.string(),
+  description: z.string(),
+});
+
+export type NewPipelineFormValues = z.infer<typeof newPipelineFormSchema>;
