+++ apps/app/src/store/sidebarStore/sidebarView.ts
@@ -1,0 +1,6 @@
+export const SidebarView = {
+  Main: "main",
+  Pipeline: "pipeline",
+} as const;
+
+export type SidebarViewType = (typeof SidebarView)[keyof typeof SidebarView];
